requirements.txt
pytest==8.4.2
python-dotenv==1.1.1
requests==2.32.5
selenium==4.36.0
faker==37.11.0
jsonschema==4.25.1
allure-pytest==2.15.0
assertpy==1.1
webdriver-manager==4.0.2
